Move app classes in separate folders
