+++ TangoRosStreamer/app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         <receiver
             android:enabled="true"
             android:exported="true"
-            android:name="eu.intermodalics.tango_ros_streamer.RunAppOnStartupBroadcastReceiver"
+            android:name="eu.intermodalics.tango_ros_streamer.android.RunAppOnStartupBroadcastReceiver"
             android:permission="android.permission.RECEIVE_BOOT_COMPLETED">
 
             <intent-filter>
@@ -38,7 +38,7 @@
         </receiver>
 
         <activity
-            android:name="eu.intermodalics.tango_ros_streamer.RunningActivity"
+            android:name="eu.intermodalics.tango_ros_streamer.activities.RunningActivity"
             android:label="@string/app_name"
             android:configChanges="orientation|screenSize|uiMode">
             <intent-filter>
@@ -48,12 +48,12 @@
         </activity>
 
         <activity
-            android:name="eu.intermodalics.tango_ros_streamer.SettingsActivity"
+            android:name="eu.intermodalics.tango_ros_streamer.activities.SettingsActivity"
             android:label="@string/title_activity_settings">
         </activity>
 
         <activity
-            android:name="eu.intermodalics.tango_ros_streamer.AboutActivity"
+            android:name="eu.intermodalics.tango_ros_streamer.activities.AboutActivity"
             android:label="@string/title_activity_about">
         </activity>
