+++ TangoRosStreamer/app/src/main/java/eu/intermodalics/tango_ros_streamer/activities/AppCompatPreferenceActivity.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.intermodalics.tango_ros_streamer;
+package eu.intermodalics.tango_ros_streamer.activities;
 import android.content.res.Configuration;
 import android.os.Bundle;
 import android.preference.PreferenceActivity;
