+++ TangoRosStreamer/app/src/main/java/eu/intermodalics/tango_ros_streamer/android/SaveMapDialog.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package eu.intermodalics.tango_ros_streamer;
+package eu.intermodalics.tango_ros_streamer.android;
 
 import android.app.Activity;
 import android.app.DialogFragment;
@@ -24,6 +24,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.EditText;
+
+import eu.intermodalics.tango_ros_streamer.R;
 
 
 /**
