+++ TangoRosStreamer/app/src/main/java/eu/intermodalics/tango_ros_streamer/nodes/ParameterNode.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package eu.intermodalics.tango_ros_streamer;
+package eu.intermodalics.tango_ros_streamer.nodes;
 
 import android.app.Activity;
 import android.content.SharedPreferences;
